Converted all frag references to int32 (Special Collision makes a reference to a large uint32 value or a negative int32 value, and also there are sometimes more fragments in a wld file than int16)
Added fragID and fragID references by each wce fragment (FragRefs). Added a map of all the fragments with the hop of using it to assign data to wce folders better.

Rewrote sprite_2d to support pitches->headings instead of just headings. This is how it is structured in wld-cli and wldcom.

Fixed write to binary issue for bminfo fragments for gequip and lgequip.

Added new way to check for variation materials. Only "_chr.wld" contain variation materials, and added fragment reference for the original of variant.